mcp/snowflake-mcp-pure/requirements.txt
mcp[cli]==1.10.1
snowflake-connector-python
python-dotenv
pydantic
